Change SUPlainInstaller to allow downgrade when versions are canary or prototype.
The approach this PR does is to check in Sparkle for the BCNYReleaseType
of the app. If it is Dia and its type is prototype or canary, it allows
the app to update even if it is a downgrade.

A more elaborate check happens before, in Dia, through the
DowngradableVersionComparison class. The code that this PR bypasses is
only a final validation performed by Sparkle to protect it from bad
actors.

My initial attempt was to set an API, but the SUPlainInstaller is
invoked in a different process on a different embedded app
(AutoUpdater), so any information I need to pass at runtime from Dia
would need to go through a very involved change in how Sparkle
is instantiated.
My second attempt was to add a new field "SUAllowDowngrade" to the
Info.plist, but then this poses a security risk, since the Info.plist is
easy to change by any user, and downgrades are considered a security
risk.

Finally, third time is a charm: I decided to make this change more
scoped. It adds knowledge about Dia and its release types to
Sparkle, which is not a good design, but I think it is a good compromise
between security, maintenance costs and pragmatism/time to develop.

Contributes to: BRSR-1004

TODO: remove the check that allows PR app type to also support
downgrade. This was added only for testing.